requirements.txt
boto3==1.9.217
botocore==1.12.217
cachetools==3.1.1
docutils==0.15.2
jmespath==0.9.4
Pillow==6.1.0
psycopg2==2.7.7
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.25.3
